% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{clean_xtab_names_transform}
\alias{clean_xtab_names_transform}
\title{transform function for get_sqlite_xtab cleaning names}
\usage{
clean_xtab_names_transform(data)
}
\description{
transform function for get_sqlite_xtab cleaning names
}
