% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{close_sqlite_xtab}
\alias{close_sqlite_xtab}
\title{close_sqlite_xtab}
\usage{
close_sqlite_xtab(xtab)
}
\arguments{
\item{xtab}{connection to the xtab as retured from get_sqlite_xtab}
}
\description{
close_sqlite_xtab
}
