% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{connect_and_normalize}
\alias{connect_and_normalize}
\title{Collect data and normalize result}
\usage{
connect_and_normalize(connection, disconnect = FALSE)
}
\arguments{
\item{connection}{SQlite connection}

\item{disconnect}{Optional parameter to determine if the SQlite connection should be closed, `FALSE` by default.}
}
\value{
A tibble with the data in long form
}
\description{
Collect data and normalize result
}
