% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{data2sqlite}
\alias{data2sqlite}
\title{convert tibble or data frame to sqlite}
\usage{
data2sqlite(data, sqlite_file, table_name, append = FALSE)
}
\arguments{
\item{data}{input tibble or data frame}

\item{sqlite_file}{output sql database path}

\item{table_name}{sql table name}

\item{append}{optional parameter, append to database or overwrite, defaul=`FALSE`}
}
\description{
convert tibble or data frame to sqlite
}
