% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{gather_for_grep_string}
\alias{gather_for_grep_string}
\title{Pivot xtab columns to long form}
\usage{
gather_for_grep_string(data, gather_key, grep_string)
}
\arguments{
\item{data}{input tibble}

\item{gather_key}{name of variable column when pivoting to long form}

\item{grep_string}{string used to identify column names to pivot by regexp match}
}
\description{
Pivot xtab columns to long form
}
