% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{gather_xtab_values_transform}
\alias{gather_xtab_values_transform}
\title{transform function for get_sqlite_xtab cleaning names}
\usage{
gather_xtab_values_transform(data, remove_member_id = TRUE)
}
\description{
transform function for get_sqlite_xtab cleaning names
}
