% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{get_sqlite_xtab}
\alias{get_sqlite_xtab}
\title{get_sqlite_xtab}
\usage{
get_sqlite_xtab("98-400-X2016288","https://www12.statcan.gc.ca/census-recensement/2016/dp-pd/dt-td/CompDataDownload.cfm?LANG=E&PID=111849&OFT=CSV")
}
\arguments{
\item{code}{Code for the xtab, used for local caching}

\item{url}{url to download the csv version of the xtab}

\item{cache_dir}{optional cache directory.}

\item{existing_unzip_path}{optional path to unzipped xtab, useful for large (>4GB unzipped) xtabs on windows machines}
}
\description{
get_sqlite_xtab
}
