% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{index_xtab_fields}
\alias{index_xtab_fields}
\title{create indices on index_fields in the sqlite database in case they don't exist}
\usage{
index_xtab_fields(sqlite_path, table_name, index_fields = NULL)
}
\arguments{
\item{sqlite_path}{path to the sqlite database}

\item{table_name}{name of the table}

\item{index_field}{optional vector of column names to index, defaul indexes all columns}
}
\description{
create indices on index_fields in the sqlite database in case they don't exist
}
