% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{strip_columns_for_grep_string}
\alias{strip_columns_for_grep_string}
\title{remove columns by regular expression}
\usage{
strip_columns_for_grep_string(data, grep_string)
}
\arguments{
\item{data}{input data frame}

\item{grep_string}{regexp to match against data frame names, matches will be removed}
}
\description{
remove columns by regular expression
}
