% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{xml_to_csv}
\alias{xml_to_csv}
\title{Use python to parse xml}
\usage{
xml_to_csv(
  code,
  url,
  python_path = "/opt/anaconda3/bin/python3.7",
  refresh = FALSE,
  temp = NA,
  conda = "/opt/anaconda3/bin/conda"
)
}
\arguments{
\item{code}{xtab identifier}

\item{url}{download url}

\item{python_path}{path to local python}

\item{refresh}{optionally redownload and reprocess the xtab, default \code{FALSE}}

\item{temp}{optionally specify path of downloaded zip file}

\item{conda}{path to local conda}
}
\description{
Use python to parse xml
}
