% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{xml_via_python}
\alias{xml_via_python}
\title{Download xml xtab data from url, tag with code for caching
useful for older (2006) data that does not come as csv option}
\usage{
xml_via_python(
  code,
  url,
  python_path = "/opt/anaconda3/bin/python3.7",
  conda = "/opt/anaconda3/bin/conda",
  refresh = FALSE,
  temp = NA
)
}
\arguments{
\item{code}{The statcan table code, needed for caching and for locating the table to extract}

\item{url}{The url to the xml table download. Only needed if the table is not cached, but
reommended to always include for reproducibility}

\item{python_path}{path to python executable, expects a conda environment 'r-reticulate' to exist.
Can be set up with `install_python_environment`.}

\item{refresh}{Will refresh cached data if set to TRUE, default is FALSE}

\item{temp}{A path to the downloaded xml zip file. Useful if file has already been downloaded and should not be
downloaded again.}
}
\description{
Download xml xtab data from url, tag with code for caching
useful for older (2006) data that does not come as csv option
}
