% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{xml_xtab_for}
\alias{xml_xtab_for}
\title{Download xml xtab data from url, tag with code for caching
useful for older (2006) data that does not come as csv option}
\usage{
xml_xtab_for(
  code,
  url,
  refresh = FALSE,
  time_value = NA,
  temp = NA,
  system_unzip = FALSE
)
}
\arguments{
\item{code}{The statcan table code, needed for caching and for locating the table to extract}

\item{url}{The url to the xml table download. Only needed if the table is not cached, but
reommended to always include for reproducibility}

\item{refresh}{Will refresh cached data if set to TRUE, default is FALSE}

\item{time_value}{If set, will use this as the time variable, otherwise extract the time variable from the data.
Setting this will slightly speed up the parsing.}

\item{temp}{A path to the downloaded xml zip file. Useful if file has already been downloaded and should not be
downloaded again.}
}
\description{
Download xml xtab data from url, tag with code for caching
useful for older (2006) data that does not come as csv option
}
