% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{get_sqlite_xtab}
\alias{get_sqlite_xtab}
\title{get_sqlite_xtab}
\usage{
get_sqlite_xtab(
  code,
  url = NA,
  cache_dir = getOption("statcanXtab.cache_dir"),
  transform = standard_xtab_transform,
  refresh = FALSE,
  existing_unzip_path = NA,
  format = "csv"
)
}
\arguments{
\item{code}{Code for the xtab, used for local caching}

\item{url}{url to download the csv version of the xtab}

\item{cache_dir}{optional cache directory, default is \code{getOption("statcanXtab.cache_dir")}.}

\item{transform}{transform to apply to data, default is \code{standard_xtab_transform} which removes the notes
column, cleans up the variable names and casts the \code{Value} column to numeric.}

\item{refresh}{redownload and parse data, default \code{FALSE}}

\item{existing_unzip_path}{optional path to unzipped xtab, useful for large (>4GB unzipped) xtabs on windows machines}

\item{format}{data format of xtab, defaults to "csv". Other valid options are "xml" and "python_xml". Set according
to the format of the data, and if conversion through xml should be done via python for better performance}
}
\description{
get_sqlite_xtab
}
