% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{install_python_environment}
\alias{install_python_environment}
\title{Set up virtual conda env}
\usage{
install_python_environment(
  python_path = "/opt/anaconda3/bin/python3.7",
  conda = "/opt/anaconda3/bin/conda"
)
}
\arguments{
\item{python_path}{path to local python}

\item{conda}{path to local conda}
}
\description{
Set up virtual conda env
}
