% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{install_python_environment}
\alias{install_python_environment}
\title{Set up virtual conda env}
\usage{
install_python_environment(python_path = "/anaconda3/bin/python3")
}
\description{
Set up virtual conda env
}
