% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{numeric_xtab_values_transfrom}
\alias{numeric_xtab_values_transfrom}
\title{transform function for get_sqlite_xtab cleaning names}
\usage{
numeric_xtab_values_transfrom(data)
}
\arguments{
\item{data}{input data frame}
}
\description{
transform function for get_sqlite_xtab cleaning names
}
