% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{standard_xtab_transform}
\alias{standard_xtab_transform}
\title{transform function for get_sqlite_xtab cleaning names}
\usage{
standard_xtab_transform(data, remove_member_id = TRUE)
}
\description{
transform function for get_sqlite_xtab cleaning names
}
