% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{tidy_xtab}
\alias{tidy_xtab}
\title{transforms standard StatCan xtab to long form}
\usage{
tidy_xtab(data, remove_member_id = FALSE)
}
\arguments{
\item{data}{data from get_sqlite_xtab afer calling collect()}
}
\description{
transforms standard StatCan xtab to long form
}
