% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{xml_to_csv}
\alias{xml_to_csv}
\title{Use python to parse xml}
\usage{
xml_to_csv(
  code,
  url,
  python_path = "/opt/anaconda3/bin/python3.7",
  refresh = FALSE,
  time_value = NA,
  temp = NA,
  conda = "/opt/anaconda3/bin/conda"
)
}
\description{
Use python to parse xml
}
